Merging feature/security policy to develop

+++ app/Http/Controllers/UserController.php
@@ -5,24 +5,36 @@ namespace App\Http\Controllers;
 use App\Http\Requests\UserRequest;
 use App\Http\Requests\UserUpdateRequest;
 use App\Models\User;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
 {
+    use AuthorizesRequests;
+
     public function index() : JsonResponse {
+        
+        $this->authorize('viewAny', User::class);
+
         $users = User::all();
         
         return response()->json(['data' => $users], 200);
     }
 
     public function find(string $id) : JsonResponse {
+
         $user = User::findOrFail($id);
+
+        $this->authorize('view', $user);
         
         return response()->json(['data' => $user], 200);
     }
 
     public function store(UserRequest $request) : JsonResponse {
+
+        $this->authorize('create', User::class);
+
         $user = User::create([
             'name' => $request->name,
             'email' => $request->email,
@@ -35,6 +47,8 @@ class UserController extends Controller
     public function update(UserUpdateRequest $request, string $id) : JsonResponse {
 
         $user = User::findOrFail($id);
+
+        $this->authorize('update', $user);
 
         $data = $request->only(['name', 'email', 'password']);
 
@@ -56,14 +70,15 @@ class UserController extends Controller
     }
 
     public function destroyAllUsers() : JsonResponse {
-        User::query()->delete();
-        
-        return response()->json([], 200);
+        abort(403);
     }
 
     public function destroyById(string $id) : JsonResponse {
+
         $user = User::findOrFail($id);
         
+        $this->authorize('delete', $user);
+
         $user->delete();
         
         return response()->json([], 200);
